requirements.txt
Pillow
numpy
pandas
pyarrow
fastparquet

